Any reason not to use config.console_encoding ?
Using sys.std(out|err).encoding still causes problems when output is piped. Depending on how the next process handle the pipe, .encoding can be None.

If there's a particular reason for these to use sys.stderr.encoding, that'd be nice to use a config variable, and test for its non-"Noneness", as we currently do for config.console_encoding

https://mediawiki.org/wiki/Special:Code/pywikipedia/6463