[ #2895284 ] Port of the User() class. Contributed by stanlekub. Superfluous whitespace removed by siebrand. Untested.
Comments from patch submit:
[..] this creates a class User(), which is a subclass of Page, with additional methods.

The patch adds:
* the class User() in page.py
* 2 new exceptions in exceptions.py
* an import statement for User() in __init__.py
* a minimal usrlib.py file is also provided for backwards-compatibility, in the same manner it was done for catlib.

This implementation should be fully compatible with the "old" one... with all of its defaults included. There is for example a strange discrepancy between the format of timestamps as returned by .contributions() and .loadedImages().

The methods .block() and .unblock() aren't implemented yet, and the 'autoblock' thingy should probably be tweeked a little more.

Comments on this are really welcomed [..]

https://mediawiki.org/wiki/Special:Code/pywikipedia/7716