Support to add a reference with multiple claims.
site.editSource() accepts now either a single claim
(=backward-compatible) for a single claim reference
or a list of claims for a multi claim reference

page.addSources() is new and takes a list of claims
in contrast to page.addSource()

Claim.sources is a list of dicts now. Thus, it is possible
to check if a property is contained.

Change-Id: Iac583043531ba3f30985244ffad2aef2ad549f51