Implement support for time datatype
WbTime datatype resembles the Wikibase time format with precision, before, after, and calendar model.
The precision can be set as integer (0-14) or with a human readable string, e.g., 'hour'.
If the precision is not set it is determined automatically, for example: WbTime(2013, 8) is August, 2013, precision month.

See these diffs as use case:
http://www.wikidata.org/w/index.php?title=Q150174&diff=67022161&oldid=67022124
http://www.wikidata.org/w/index.php?title=Q150174&diff=67022199&oldid=67022161

Change-Id: I4e13f797d7979d5944fa8822ea09076448b9990e
Original-Change-Id: Iac583043531ba3f30985244ffad2aef2ad549f51