Fix string formatting in copyAndKeep.
According to Python docs, section 5.6.2, "When the right argument is a dictionary
(or other mapping type), then the formats in the string must include a parenthesised
mapping key into that dictionary inserted immediately after the '%' character.
The mapping key selects the value to be formatted from the mapping."

This is not the case here.

Change-Id: I3ee69a1d44b4109fd98837be12af39a23eba2ae3