(bug 62126) Retry creating a new page at wikidata.
Bug 46535 reports that bot fails while creating a page with
error "Could not create a new page. It already exists." But
the page does not exist.

This is a temporary patch to solve this problem above until
46535 has been fixed. It tries again until the page has been
created or maxretries exceeded.

There is a remaining problem with this bug/patch. There is no way
to check whether the page does really exist or the API gives
a wrong message.

Change-Id: I30d596d814caf3bf812b38b975019aef5213c787