Add Page.revisions, getCreator, getLatestEditors
Adds Page.getCreator() and Page.getLatestEditors() to be the same as the
original compat methods. They were just added for compatibility but are
deprecated. The newly added Page property 'oldest_revision' now returns
a Revision object.

Page.revisions() is a generator to retrieve Revision objects and similar
to Page.getVersionHistory().

Bug: T75023
Change-Id: Ifee5de37e6feb7855550c2d8c9eeebc70ac542da