Add Proofread Page preloading functions
Implement preload param in API:Properties module "info" to
preload text.

Application: on Wikisource, with Extension:Proofread Page, it is
possible to preload text from a not existing page in the Page
namespace, if the corresponding Index page is present.

If a djvu file is connected to an Index page, text is basically already
available, making this script less important.

See also Bug 58963 - prop=info&inprop=preload does not return text
and Bug 64853 - port djvutext.py to core

Change-Id: Ic534597fe42c09e8432300a19e77c23a21f21a8e