(bugfix) compare the default key in the right order
if default key exist and answer is an empty string, the default
value must be returned. Return the choice if it is in hotkeys.
Otherwise try again.

With this patch the keys may be given as list or tuple or also
as string which simplifies the code. Because '' in 'anystring'
is always True the old order didn't work for hotkeys string.

Change-Id: I2caddf213818de231a3258b3a00cca0f6e51b268