harvest_template.py perf: dont parse and discard
Currently the wikipage is parsed even when all of the parsed values
will be discarded.  This is slow and noisy.

When an item has claims for all properties to be processed,
skip the item before fetching and parsing the wikitext.

Improve existing claim merge TODO to properly describe the complexity.

Change-Id: Ib3df4cce7125591e8fa585d18988e2a605456a09