(perf) Parse ItemPage title once
ItemPage may be constructed without a page title, as it is
delay loaded from the wikibase_item of a page on another site.

Optimise ItemPage.title() so it does not reparse the Link title
every invocation.

Add test case for instantiating an empty ItemPage and setting .id to
a valid Qxx, and another case for altering .id highlighting a bug
yet to be fixed.

Split from: If7ca06adbb4b9932bba0abffc7588afcb320e934

Change-Id: I95a053a1ea17611498fcb5601c09a83d82f91fe8