(bug 66466) Exception getting edit token for Claim
Site.editSource and other Claim modification methods fetched an edit token
using the claim itself as 'page', which was previously the Property:Pxx page.

Due to Id63969c045ac63a5b0276f696ce5797c4ee8b6df, Claim objects are no longer
a Page, and these methods need to select a page for which to fetch the token.

Standard practise is to use 'Main Page' to fetch an edit token, which is
what this changeset does.

Change-Id: Ic7f8308438c910e5b4ce798a36a1e6e5d77c4a96