Do not remove imported packages
Imported modules were hidden in 2009, as part of a large cleanup of the
symbols exported by config.py
https://mediawiki.org/wiki/Special:Code/pywikipedia/4285

While there is benefit in hiding symbols not intended to be public,
hiding the imported modules is not beneficial as it is unlikely that
any config value will be of type module, so they can be differentiated,
and indeed they were already skipped in 'list of config options' code.
https://mediawiki.org/wiki/Special:Code/pywikipedia/4887

The removal of symbols __sys and os currently cause pyflakes error F821.
While this isnt a functional problem, it does prevent reflection and
inspection of the loaded python code, and is unecessary.

Also moved platform import into the only code block that uses it,
meaning it wont be loaded on many systems.

Also removed 're' import, as it was unused.

Change-Id: I783fba623a2825af3273ecad286d47e897c2a674