Add Site login status to cache key
Some API responses differ depending on the privileges of the user,
which can range from logged out anon to sysop.  The cache needs to
include the login status in the key, so the cached response is not
reused for queries from a different privilege level.

For example, when the cache is empty and the user has never logged
into pywikibot, the unit tests cache site.userinfo with the details
of the logged out user, causing tests which require a logged-in user
to prompt for a password on the console many times per test, and the
site.userinfo is never updated.  With this change, only one password
prompt appears and the result is cached.

Change-Id: I5e480be86a0b7b6a5907cf71690ba9ebed0b04b6