Updated pywikibot/core Project: pywikibot/externals/httplib2  5dc145682eb726bd54eaf7e9f940844ca234db71
Rewrite python version dependent import

The top level of this external includes a __init__.py so that
it may be imported with only 'externals' in sys.path.

However it copies the contents of the python version dependent httplib2
code, resulting in module level variables appearing in two different
namespaces.  As a result, regarding bug 66161, the 'httplib2.debuglevel'
modified in pywikibot code is a different variable to the
'httplib2.debuglevel' used by the httplib2 module.

Instead of copying the python version dependent httplib2, re-import the
python version dependent httplib2.

Change-Id: Ic520505545a5f50f669a01375b253426ecad15ed