Log package versions
The version of major dependencies 'httplib2' and 'mwparserfromhell'
shoud be written to the log, and included in the script version.py.

In verbose mode it is helpful to see what other major packages are
being automatically loaded into the python environment, and their
versions.

Move the online startup checks after the local checks.

Change-Id: Iae9f73b23ab1eb1a4acce46a690c0b4f9f9396a8