textlib InterwikiFormat() depends on config
InterwikiFormat implicitly uses Page.title() which by default uses
config settings to dynamically determine whether to include a family and/or
lang prefix to the title.  InterwikiFormat is supposed to format
interwikis relative to 'insite', not config variables such as family.

Explicitly invoke Page.title() with appropriate parameters to override
config settings.

The relevant test failed when config.family was not wikipedia.
In addition to the bug in textlib, the test also depended on config
settings.  Link()'s should be instantiated with a source in these tests.

Change-Id: I9ab8780e659517db7068d4af3164a2f54ce85461