Use test.wikidata as travis default site
Unit tests should not depend on config default site in order to
function correctly.

This change ensures that any tests which do depend on the config
default site are identified and fixed quickly.

The default needs to be a WMF wiki, and wikidata is the most different
to the other WMF families as namespace 0 is not wikitext, which should
help break tests which are assuming a default site of Wikipedia.

Change-Id: Iad16d81f5e993ee320a4b56f670baf5649d16f93