Add Namespace class
Replaces the current implicit data structure of
site._namespaces with a class with variables with
names that explain their value, and methods
that simplify using namespaces.

The class also stores other attributes provided by
API siprop=namespaces.

This changeset does not implement new Namespace semantics
in other parts of pywikibot so-as to ensure that existing
usage of site._namespaces is supported by the Namespace
class.

Change-Id: I0cce21e6161031861c0056dc5498f47bc26e1cc8