Print test duration greater than 10s
unittest only provides test duration for all tests.
It is useful to see timing information per test, if
it is longer than a reasonable time period for a test,
and recording this information into travis build logs.

Tests which take a long period of time may need to be refactored to
achieve the same goal with less setup work done inside the test, or
even split into multiple tests.

Change-Id: I8134a99c37b0435ecc20a2ff571e77347f9c785e