Remove unnecessary commented out code re eo.wp
Since 2010, the HTML &#39; (') is ignored for all sites
with a comment that indicates mediawiki bug 24093 is the reason.

Before that, eo.wp had &#39; on its ignore list with
a comment indicating this diff was the reason:
https://eo.wikipedia.org/w/index.php?title=Liberec&diff=next&oldid=2320801

When this hex code was added to the global ignore list, the eo.wp
rule was commented out.
https://mediawiki.org/wiki/Special:Code/pywikipedia/8314

This changeset removes the commented out old rule, as it is
unlikely to be reinstated or needed.  Documentation of these
rules should be on the wiki, and test cases should exist to
ensure site-specific issues raised in the past are not regressed.

Change-Id: Idec1eb9af527d6a812631dd13a9b1fcf151f03f6