Remove site NotImplementedYet methods
These methods are largely copied from compat and rely on attributes
and data files which do not exist.

They have been unneeded for years and have only seen maintenance edits
to 'fix' the structure of the code as other methods change, pepifying, etc.

If the algorithms are needed again, they can be found in compat.

Change-Id: Iad18a29b9b02bdb4164b3d810fe3dd482280765b