Remove warning from docstring in blockpageschecker
The warning is not valid as the necessary 'preferences' are
not user preferences, but are per wiki configuration, and
that configuration is in place for six wikis which are
listed in variable 'project_inserted'.  A warning is given
when the script is run with a different language code.

The IRC channel name is out of date, and directions on where
to seek help should be provided for the entire package rather than
in an individual script docstring.

Change-Id: I2a76b4e813ba3bc5d9cb00a89075a4f095c586fa