Fix wikibase tests broken by wiki edit
wikibase:Q5296 English label has changed from 'Main Page'
to 'Wikimedia main page'.  Slightly adjust the logic so that either
is acceptable.

The initial version of the wikibase tests included a minor
code issue causing a test to be ineffective.  In testWikibase, instead
of exercising item2.labels, item.labels is exercised a second time.
item2.getID does not cause get() to be called, so get() needs to
be explicitly in the test sequence.

Change-Id: I75fa36c3b05ba8141c51ce19575b7667f408fd84