Detect attempt to fetch wikibase item of an entity
While the wikidata and wikispecies families dont allow transcluded data,
the current code tries to find a wikibase item for wikidata pages, which
causes a warning like the following and raises an APIError exception.
    WARNING: API warning (wbgetentities):
    Unrecognized value for parameter 'sites': testwikidatawiki

This change detects whn ItemPage.fromPage is being asked to load an item
for a page on a site that cant have an item, and raises WikiBaseError
instead of APIError.

Change-Id: Ie89289536a777b8f6bb92b09d544c13cf32f5df4