update Wikibase item on category rename
The 'wikibase' variable in CategoryMoveRobot is True by default,
but can be turned off by passing the '-nowb' argument.

If the bot cannot move categories (thus updating the item
automatically) and no registered user can be found for the
Wikibase repository, pywikibot.NoUsername is raised
and caught in main().

other unrelated parts of the code have not and will not be touched

bug: 64815
Change-Id: Idaaa272b30ad2e3372e70e0c6af4995998546d2d