Wikibase test fix
The wikibase tests introduced recently are now failing.
They were introduced here:
Ic0e5ffd7526d71ca8d460f29ed37747a0b255930

That change passed all tests; see here:
https://travis-ci.org/wikimedia/pywikibot-core/builds/30325106

That log shows that it introduces a new warning like the following
and raised an APIError exception on an attempt to load an item.
    WARNING: API warning (wbgetentities):
    Unrecognized value for parameter 'sites': testwikidatawiki

That warning appears until travis build 1078, and then appears in 1079
https://travis-ci.org/wikimedia/pywikibot-core/jobs/30562987
https://travis-ci.org/wikimedia/pywikibot-core/jobs/30661149

It appears to be a server config change that has fixed that
API warning and APIError exception, and now the exception is NoPage.

Change-Id: I3d1655b53bb11236545502634dad667ecef74d53