Remove custom lunatic-python patches
Replace custom instructions in externals/README and remove custom
patch file externals/patch-lua as several github repos include that
change are a few are advanced well past that patch.

If a custom lua is required, it should be placed in externals as
'lua' so the pywikibot code finds it automatically without looking
for '_lua'.

Change-Id: I096e55da10323b1a77fcc9f822f77ba23610c563