Script test fixes for py2.6
unittest2 on py2.6 on travis reports the following warning and
returns a non-zero exit code for the test suite when a test
is unexpectedly successful.

DeprecationWarning: Use of a TestResult without an addUnexpectedSuccess
method is deprecated

Several of our tests fail only on one site (en.wp, ar.wp or wd), or on
other conditions which cant be predicted.

This test removes use of unexpected success from the tests.

Change-Id: Ic2a4be02871f9eda2c40744935392985cd7a6a01