Perform site tests against the config site
Add WikimediaDefaultSiteTestCase to specify the test class site
should be a WMF site, using the default site if it is a WMF site.

Run the site tests against the config defined site, or
a WMF site as required.

The travis build matrix will catch differences on each
defined site, for each platform.

Change-Id: I466f87f3d7db30ccfa75f6425c2383822a1ae307