new Bot.current_page property to set and log the working page
Nearly every script that 'treats' a page used to print its title
on the screen, sometimes in purple, sometimes prepended with
"Processing..." or "Working on...".
Now the setter uses the standard output for the colorized message
and the logging.VERBOSE level for the %r-formatted string,
improving consistency and semantic value of the console output.

This property also allows to get the current working page
at any time, and to avoid printing the same title twice.

Change-Id: I886c48c7bf7b66917be814df3833f70caa22bd4a