Allow travis failures on python 2.6
Python 2.6 setup is registering a fail on travis builds
due to SkipTest occurring in setUpClass.  The same SkipTest
is registered as a skipped test on python 2.7 and doesnt
cause the build to be considered a failure.

https://travis-ci.org/wikimedia/pywikibot-core/builds/31882466

travis allows builds to fail without the changeset being considered
a failure.  The builds allowed to fail are displayed as red, with
the log to allow issues to be investigated and fixed without putting
a halt on merging of changesets.

Change-Id: I0db9294154f06d650928b144390b383144d87d91