Deprecate SSL override configuration
No sites use this functionality, as wikimedia sites changed from
a https://secure.wikimedia.org front-end processor to using https
on the target wiki.

Also remove the 'ssl' override parameter from http.request.
The ssl param was only used with the site parameter, to support the
configuration variables, and was ignored if a uri was provided.

Change-Id: I5834ca17972e888886d15524098331cd8839f1da