Bug 69315 - cs timestamp not supported (II)
New fix for bug in cs wiki.

If months are expressed as digits, regex expression for months
needs to used \d{1,2} and not the names as strings.
This ensure proper match of digits.
In such cases, it is assumed that date is expressed as dd. mm.

When only two occurrences of month (cnt=2), leave them to allow
to search for dd. mm. pattern.

Change-Id: Ic8549b2d5502fca86b3bb87e92001949f78720b0