PropertyPage.get() fails
PropertyPage method get attempts to set 'type', which is a
property of superclass Property that does not have a setter, resulting
in an AttributeError:

    self.type = self._content['datatype']
AttributeError: can't set attribute

Fix by putting the datatype into the underlying _type attribute.

Also use the 'type' property in PropertyPage.newClaim so that each newly
created Claim does not need to fetch the property type.

Change-Id: I845ec3ad91017e386e568ee5bcb85fd48644adfb