Fix backwards compatibility of APISite.hasExtension()
Previously that method used "name" and "unknown", but "unknown" was
removed in d42d519adf59b92998ba7b87008c58b7193a7014 so tests broke
which used it as a positional parameter.

This adds a method has_extension which only accepts one parameter
while the old hasExtension supports both (but ignores the second
completely).

This also fixes the site tests which may relied on it so various
tests aren't applicable anymore. There is only one test which makes
sense but is deactived, because it won't work unless it is possible
to make "need_version" believe that the current version doesn't
support it.

Change-Id: I22a17b0f9eca208daaa3b3eeeab1b86b5c34b399