[FEAT] Use the "default"-field in Siteinfo to store cache time
This allows Siteinfo to automatically reload after a certain time
has expired. It is also possible to determine the time of the last
request time for each field. As a use case live_version now caches
it for up to a day

The cache time is the time of the request, but the request might be
cached separately (which is the case in the tests) so it's not the
time the server actually send it.

Default values in the caches (siprops which the server didn't support)
are always expired and are always rerequested, except when the expiry
is set to False (that value basically never expires).

Change-Id: Iff4b6a68e4cdf7f3fb5be17d6704d2b23cae2175