[FEAT] Add deprecation warnings for textlib/deprecation redirects
It's adding a function which dynamically can create a redirect to
another function and warn because of deprecation when the
redirected function is executed. It works similar to the
@deprecated decorator but doesn't require the old function.

All functions which from 'textlib' which were simply redirected
from 'pywikibot' to the 'textlib' module are created with this
function. Also 'deprecated' and 'deprecate_arg' are marked as
deprecated too.

Change-Id: I69dbf351c22fd0caacbf047ae38aca067f8cef96