[FEAT] Handle API warnings
Some code parses the warnings returned by the API separately like
Siteinfo. This option adds the possibility to handle warnings
returned by the api via "_warning_handler". If it's not a method
or that method returns not True the warning is still output like
before (and is by default). The method must accept two parameters
and is executed for each warning.

This also shows warnings when CachedRequest are used and not only
for the first actually submitted request.

Change-Id: I62d13dca5b592bddf72dbbdcc0b8edf77c0059ed