Move deprecator methods to tools module
The methods deprecated and deprecate_arg are not needed by all
pywikibot usage, so do not belong in __init__.

As they are quite generic, except for the need to present a
message to the user, they can reside in the tools module.
The methods have been moved there without modification, and
two variables (warning and debug) added to the tools module
so they behave the same.  There is a slight improvement, that
they no longer have a hard dependency on the bot logging layer
and therefore may be used earlier in the initialisation process
or in custom code without importing the pywikibot package.

Change-Id: I1a7520617d11ed18e541b0e4bce28a775c3f131a