upload.py: Add option to automatically abort upload on warning
Useful when a script is issuing a batch of upload.py commands to upload
a lot of files, and we want to skip files that are already uploaded
without user intervention. Unfortunately, it will also skip the upload
if there are warnings not related to upload over an existing filename,
but because how warnings are handled, there's no way to tell what kind
of warning we've received.

Bug: 69852
Change-Id: I5e23049fb51f1804c230bff67525baa15e4c1f1c