Use ipaddress module if installed
page.py includes a regexp to check if a username is an IPv4 or IPv6
address, for User.isAnonymous().

There are other instances where pywikibot should be validating strings
which may contain a IP address, such as config parsing, and a client
side assert to prevent writing to the server as an anonymous user.

Python 3 includes a ipaddress module, and there are functional backports
available.

Two of the IP6 tests fail with the Python 3 ipaddress module and backport,
which is http://bugs.python.org/issue22282 .  These are only corner cases,
unlikely to be a serious problem.

If the ipaddress exists and is functional, use it, otherwise
fall back to the old regex approach.

Bug: T76286
Change-Id: I438632608c07f2b7ed196e8b2a536481fcfa3e8e