family.py: make Family() comparable with strings
Replace Family.__cmp__() with __eq__() and __ne__().
Now it will compare also to an object which is not an instance
of Family() and without a name attribute by trying to load the
corresponding family.

This will allow comparisons like:
site.family == pywikibot.config.family

Moved Family() from site.py to family.py and converted to staticmethod
Family.load().

Change-Id: Id3da11fd8a6e649bee768dbb42f53e7167fe43e3