[FIX] Improved flake8/flake8-docstring handling
In 5426720bfe83e37aec032e8afcd9470294293ac4 the textlib redirects
have been dynamically deprecated, but because flake8 is unable to
detect that __all__ gets concatinated a workaround was used which
'fixed' flake8. But pep257 throws an AllError as soon as __all__
appears somewhere else. pep257 also throws an AllError, when the
content of the list doesn't begin in the same line as the __all__.

This also fixes the warning pep257 returned, when __all__ wasn't
a tuple in tests/__init__.

Change-Id: I6dc66f0082ae594c63b99c8c022ab38cdc705dec