Fix wikibase test test_own_client
The test test_own_client was introduced when Wikidata became its
own client.  The test checked whether pywikibot could support this
configuration, but it used an invalid test scenario to do this.
It was introduced in dce03581d4a2f0e52aa5b7a2e2b1e68a5a1b48a1

The Main Page should not be a valid ItemPage title; that is being
fixed in other changes.  The intention was to see whether an item
could be loaded for the normal page object for Main Page.

Change-Id: I21686a13825f547cf1a4ea34dae697139f3e0bbe