[FIX] Category: Load the category dump lazy
The script test failed, because it always told it loaded the
category dump (if the file existed). This loads the category dump
only as soon as it's used. Because that instance is now also storing
the filename it's possible to dump it into the same file without
determining the filename.

Change-Id: Ia3aa847afd18d21a5ae85e5c2647178ef29fba4a