[FIX] Don't use wikibase and move pages only if valid
If the site doesn't has a wikibase repository, it automatically sets
'-nowb'. Also document what it does.

Additionally it only moves a page if the original exists but not the
target. It warns if it couldn't move a page because the target
exists. If the original category page doesn't exists it warns too,
but only logs if the original talke page doesn't exists. It only
deletes a page if the category has been removed or if the page has
been moved.

Two options have been added:
* -mvtogether: Only change the pages' category if the category
               page has been moved.
* -allowsplit: If not set it only moves the category page and
               category talk page if both pages can be moved.

Bug: 65981 (not original, the bug later described in the comments)
Change-Id: I0cfb688ee4fdd0529af287720385807d0f0fbd81