Require login for two site methods
Also purgepages was recently added as a API 'write' action
which asserts it must be performed by a logged in user.

is_uploaddisabled added recently for bug 69090 only
works if the user is logged in, otherwise the assert
module returns an error before the upload module can
respond.  Re-raise any API error that is_uploaddisabled
is not expecting.

Use test projects to test is_uploaddisabled, as the
travis config already has username's for those sites.

Change-Id: I009260c9476852cc8ee818aa791fd15f1dc05ec1