[FEAT] Add mandatory docstring tests
Currently the docstring tests are not mandatory. Because with any
change there is a chance that new PEP 257 errors are introduced it's
hard to make them mandatory, because all errors need to be fixed at
once. This change should make it easier as particular files can be
fixed from PEP 257 errors and then added to the mandatory list. If
now any later change introduces an error, it will fail the test on
already fixed files.

This adds only the test environment for tox but not actually adds
the test, which needs to be done separately.

All files which are PEP 257 compliant (except family files) are
already added.

Change-Id: I3e34ecc56996d58ace96299ceaf63a2278055fcb