Use Pillow instead of PIL in setup.py
PIL stopped working in travis builds.
It works correctly under setuptools 5.7, but fails under setuptools 3.3
Rather than force a setuptools upgrade on everyone, use Pillow.

'PIL is not setuptools compatible.'
http://pillow.readthedocs.org/en/latest/about.html#why-a-fork

Remove the apt-get of python-imaging-tk, which previously
installed PIL.
'System Python is not used and should not be relied on.'
http://docs.travis-ci.com/user/languages/python/

Change-Id: Ic437462da8b637ce17a333c19a62b02a63474a07