Remove globals from test cases
Add TestCase methods get_mainpage and get_missing_article,
with caching in get_mainpage, to replace global variables
in the test modules.

Also move the 'Site.movepage' test to edit_failure_tests,
as it is a 'write' API call, and it may unexpectedly do a write
if there is a bug in the codebase.
After this move, if it ever does act unexpectedly,
it does it on a test wiki.

Change-Id: I94ae444d440f366b353231eecd37ebe1f5411605