Preload pywikibot library in pwb shell
The pwb shell script has not pre-loaded the pywikibot library, or
injected any other python objects into the shell environment.
This is useful, when the user needs to set up their own environment.

However the more common case is the user wants to be able to
use pywikibot command line arguments, and pre-load the pywikibot library.

Change-Id: I976a4b19d7a3eebb13a37dccce6bbe6454294c78