Raise exception on unknown MediaWiki exception
Some MediaWiki exceptions should be handled by backing off
and retrying, however others will not resolve themselves
and need to be a python exception

There was a list of three exception that caused a re-try,
but d5cc2b9 removed the list and always retried when there
was a MediaWiki exception.

This change effectively reverts d5cc2b9, but also adds
APIMWException as a subclass of APIError, and adds information
about the three exceptions which should cause an API retry.

Bug: 70754
Change-Id: I3c4d23bfc4f8feb01fc6e349d271fceb746412d8